refactor: 운동 라이프사이클 컨트롤러 분리 - 운동 생성, 삭제, 수정, 상세 조회
