+++ src/test/java/umc/cockple/demo/domain/exercise/integration/ExerciseCommandIntegrationTest.java
@@ -89,401 +89,6 @@ class ExerciseCommandIntegrationTest extends IntegrationTestBase {
     }
 
     @Nested
-    @DisplayName("POST /api/parties/{partyId}/exercises - 운동 생성")
-    class CreateExercise {
-
-        private ExerciseCreateDTO.Request validRequest;
-
-        @BeforeEach
-        void setUp() {
-            validRequest = ExerciseCreateDTO.Request.builder()
-                    .date("2099-12-31")
-                    .buildingName("테스트 체육관")
-                    .roadAddress("서울특별시 강남구 테헤란로 1")
-                    .latitude(37.5)
-                    .longitude(127.0)
-                    .startTime("10:00")
-                    .endTime("12:00")
-                    .maxCapacity(10)
-                    .allowMemberGuestsInvitation(true)
-                    .allowExternalGuests(false)
-                    .build();
-        }
-
-        @Nested
-        @DisplayName("성공 케이스")
-        class Success {
-
-            @Test
-            @DisplayName("201 - 모임장이 운동을 생성하면 exerciseId를 반환한다")
-            void owner_createExercise() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                mockMvc.perform(post("/api/parties/{partyId}/exercises", party.getId())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(validRequest)))
-                        .andExpect(status().isCreated())
-                        .andExpect(jsonPath("$.data.exerciseId").isNumber());
-            }
-
-            @Test
-            @DisplayName("201 - 부모임장도 운동을 생성할 수 있다")
-            void subManager_createExercise() throws Exception {
-                SecurityContextHelper.setAuthentication(subManager.getId(), subManager.getNickname());
-
-                mockMvc.perform(post("/api/parties/{partyId}/exercises", party.getId())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(validRequest)))
-                        .andExpect(status().isCreated())
-                        .andExpect(jsonPath("$.data.exerciseId").isNumber());
-            }
-        }
-
-        @Nested
-        @DisplayName("실패 케이스")
-        class Failure {
-
-            @Test
-            @DisplayName("404 - 존재하지 않는 파티면 에러를 반환한다")
-            void partyNotFound() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                mockMvc.perform(post("/api/parties/{partyId}/exercises", 999L)
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(validRequest)))
-                        .andExpect(status().isNotFound())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.PARTY_NOT_FOUND.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.PARTY_NOT_FOUND.getMessage()));
-            }
-
-            @Test
-            @DisplayName("404 - SecurityContext의 멤버가 DB에 없으면 에러를 반환한다")
-            void memberNotFound() throws Exception {
-                // SecurityContext에는 존재하지 않는 memberId 세팅
-                SecurityContextHelper.setAuthentication(999L, "없는멤버");
-
-                mockMvc.perform(post("/api/parties/{partyId}/exercises", party.getId())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(validRequest)))
-                        .andExpect(status().isNotFound())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.MEMBER_NOT_FOUND.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.MEMBER_NOT_FOUND.getMessage()));
-            }
-
-            @Test
-            @DisplayName("400 - 비활성화된 파티면 에러를 반환한다")
-            void inactiveParty() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                party.delete();
-                partyRepository.save(party);
-
-                mockMvc.perform(post("/api/parties/{partyId}/exercises", party.getId())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(validRequest)))
-                        .andExpect(status().isBadRequest())
-                        .andExpect(jsonPath("$.code").value(PartyErrorCode.PARTY_IS_DELETED.getCode()))
-                        .andExpect(jsonPath("$.message").value(PartyErrorCode.PARTY_IS_DELETED.getMessage()));
-            }
-
-            @Test
-            @DisplayName("403 - 일반 멤버가 생성 시 에러를 반환한다")
-            void normalMember_forbidden() throws Exception {
-                SecurityContextHelper.setAuthentication(normalMember.getId(), normalMember.getNickname());
-
-                mockMvc.perform(post("/api/parties/{partyId}/exercises", party.getId())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(validRequest)))
-                        .andExpect(status().isForbidden())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.INSUFFICIENT_PERMISSION.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.INSUFFICIENT_PERMISSION.getMessage()));
-            }
-
-            @Test
-            @DisplayName("400 - 시작 시간이 종료 시간 이후면 에러를 반환한다")
-            void invalidExerciseTime() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                ExerciseCreateDTO.Request invalidRequest = ExerciseCreateDTO.Request.builder()
-                        .date("2099-12-31")
-                        .buildingName("체육관")
-                        .roadAddress("서울특별시 강남구 테헤란로 1")
-                        .latitude(37.5).longitude(127.0)
-                        .startTime("12:00").endTime("10:00")
-                        .maxCapacity(10)
-                        .allowMemberGuestsInvitation(true)
-                        .allowExternalGuests(false)
-                        .build();
-
-                mockMvc.perform(post("/api/parties/{partyId}/exercises", party.getId())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(invalidRequest)))
-                        .andExpect(status().isBadRequest())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.INVALID_EXERCISE_TIME.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.INVALID_EXERCISE_TIME.getMessage()));
-            }
-
-            @Test
-            @DisplayName("400 - 과거 시간으로 운동 생성 시 에러를 반환한다")
-            void pastTime() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                ExerciseCreateDTO.Request pastRequest = ExerciseCreateDTO.Request.builder()
-                        .date("2000-01-01")
-                        .buildingName("체육관")
-                        .roadAddress("서울특별시 강남구 테헤란로 1")
-                        .latitude(37.5).longitude(127.0)
-                        .startTime("10:00").endTime("12:00")
-                        .maxCapacity(10)
-                        .allowMemberGuestsInvitation(true)
-                        .allowExternalGuests(false)
-                        .build();
-
-                mockMvc.perform(post("/api/parties/{partyId}/exercises", party.getId())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(pastRequest)))
-                        .andExpect(status().isBadRequest())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.PAST_TIME_NOT_ALLOWED.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.PAST_TIME_NOT_ALLOWED.getMessage()));
-            }
-        }
-    }
-
-    @Nested
-    @DisplayName("DELETE /api/exercises/{exerciseId} - 운동 삭제")
-    class DeleteExercise {
-
-        private Exercise exercise;
-
-        @BeforeEach
-        void setUp() {
-            exercise = exerciseRepository.save(
-                    ExerciseFixture.createExercise(party, LocalDate.of(2099, 12, 31)));
-        }
-
-        @Nested
-        @DisplayName("성공 케이스")
-        class Success {
-
-            @Test
-            @DisplayName("200 - 모임장이 운동을 삭제하면 deletedExerciseId를 반환한다")
-            void owner_deleteExercise() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                mockMvc.perform(delete("/api/exercises/{exerciseId}", exercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.deletedExerciseId").value(exercise.getId()));
-            }
-
-            @Test
-            @DisplayName("200 - 부모임장도 운동을 삭제할 수 있다")
-            void subManager_deleteExercise() throws Exception {
-                SecurityContextHelper.setAuthentication(subManager.getId(), subManager.getNickname());
-
-                mockMvc.perform(delete("/api/exercises/{exerciseId}", exercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.deletedExerciseId").value(exercise.getId()));
-            }
-        }
-
-        @Nested
-        @DisplayName("실패 케이스")
-        class Failure {
-
-            @Test
-            @DisplayName("404 - 존재하지 않는 운동이면 에러를 반환한다")
-            void exerciseNotFound() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                mockMvc.perform(delete("/api/exercises/{exerciseId}", 999L))
-                        .andExpect(status().isNotFound())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.EXERCISE_NOT_FOUND.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.EXERCISE_NOT_FOUND.getMessage()));
-            }
-
-            @Test
-            @DisplayName("404 - SecurityContext의 멤버가 DB에 없으면 에러를 반환한다")
-            void memberNotFound() throws Exception {
-                SecurityContextHelper.setAuthentication(999L, "없는멤버");
-
-                mockMvc.perform(delete("/api/exercises/{exerciseId}", exercise.getId()))
-                        .andExpect(status().isNotFound())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.MEMBER_NOT_FOUND.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.MEMBER_NOT_FOUND.getMessage()));
-            }
-
-            @Test
-            @DisplayName("403 - 일반 멤버가 삭제 시 에러를 반환한다")
-            void normalMember_forbidden() throws Exception {
-                SecurityContextHelper.setAuthentication(normalMember.getId(), normalMember.getNickname());
-
-                mockMvc.perform(delete("/api/exercises/{exerciseId}", exercise.getId()))
-                        .andExpect(status().isForbidden())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.INSUFFICIENT_PERMISSION.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.INSUFFICIENT_PERMISSION.getMessage()));
-            }
-        }
-    }
-
-    @Nested
-    @DisplayName("PATCH /api/exercises/{exerciseId} - 운동 수정")
-    class UpdateExercise {
-
-        private Exercise exercise;
-        private ExerciseUpdateDTO.Request validRequest;
-
-        @BeforeEach
-        void setUp() {
-            exercise = exerciseRepository.save(
-                    ExerciseFixture.createExercise(party, LocalDate.of(2099, 12, 31)));
-
-            validRequest = new ExerciseUpdateDTO.Request(
-                    "2099-12-31",
-                    "수정된 체육관",
-                    "서울특별시 강남구 테헤란로 2",
-                    37.6,
-                    127.1,
-                    "11:00",
-                    "13:00",
-                    12,
-                    false,
-                    true,
-                    "공지사항"
-            );
-        }
-
-        @Nested
-        @DisplayName("성공 케이스")
-        class Success {
-
-            @Test
-            @DisplayName("200 - 모임장이 운동을 수정하면 exerciseId를 반환한다")
-            void owner_updateExercise() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                mockMvc.perform(patch("/api/exercises/{exerciseId}", exercise.getId())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(validRequest)))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.exerciseId").value(exercise.getId()));
-
-                Exercise updatedExercise = exerciseRepository.findById(exercise.getId()).orElseThrow();
-                assertThat(updatedExercise.getPartyGuestAccept()).isFalse();
-                assertThat(updatedExercise.getOutsideGuestAccept()).isTrue();
-            }
-
-            @Test
-            @DisplayName("200 - 부모임장도 운동을 수정할 수 있다")
-            void subManager_updateExercise() throws Exception {
-                SecurityContextHelper.setAuthentication(subManager.getId(), subManager.getNickname());
-
-                mockMvc.perform(patch("/api/exercises/{exerciseId}", exercise.getId())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(validRequest)))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.exerciseId").value(exercise.getId()));
-            }
-        }
-
-        @Nested
-        @DisplayName("실패 케이스")
-        class Failure {
-
-            @Test
-            @DisplayName("404 - 존재하지 않는 운동이면 에러를 반환한다")
-            void exerciseNotFound() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                mockMvc.perform(patch("/api/exercises/{exerciseId}", 999L)
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(validRequest)))
-                        .andExpect(status().isNotFound())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.EXERCISE_NOT_FOUND.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.EXERCISE_NOT_FOUND.getMessage()));
-            }
-
-            @Test
-            @DisplayName("404 - SecurityContext의 멤버가 DB에 없으면 에러를 반환한다")
-            void memberNotFound() throws Exception {
-                SecurityContextHelper.setAuthentication(999L, "없는멤버");
-
-                mockMvc.perform(patch("/api/exercises/{exerciseId}", exercise.getId())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(validRequest)))
-                        .andExpect(status().isNotFound())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.MEMBER_NOT_FOUND.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.MEMBER_NOT_FOUND.getMessage()));
-            }
-
-            @Test
-            @DisplayName("403 - 일반 멤버가 수정 시 에러를 반환한다")
-            void normalMember_forbidden() throws Exception {
-                SecurityContextHelper.setAuthentication(normalMember.getId(), normalMember.getNickname());
-
-                mockMvc.perform(patch("/api/exercises/{exerciseId}", exercise.getId())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(validRequest)))
-                        .andExpect(status().isForbidden())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.INSUFFICIENT_PERMISSION.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.INSUFFICIENT_PERMISSION.getMessage()));
-            }
-
-            @Test
-            @DisplayName("400 - 이미 시작된 운동이면 에러를 반환한다")
-            void alreadyStarted() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                Exercise startedExercise = exerciseRepository.save(
-                        ExerciseFixture.createExercise(party, LocalDate.of(2000, 1, 1)));
-
-                mockMvc.perform(patch("/api/exercises/{exerciseId}", startedExercise.getId())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(validRequest)))
-                        .andExpect(status().isBadRequest())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.EXERCISE_ALREADY_STARTED_UPDATE.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.EXERCISE_ALREADY_STARTED_UPDATE.getMessage()));
-            }
-
-            @Test
-            @DisplayName("400 - 시작 시간이 종료 시간 이후면 에러를 반환한다")
-            void invalidTime() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                ExerciseUpdateDTO.Request invalidTimeRequest = new ExerciseUpdateDTO.Request(
-                        "2099-12-31", null, null, null, null,
-                        "13:00", "11:00", null, null, null, null
-                );
-
-                mockMvc.perform(patch("/api/exercises/{exerciseId}", exercise.getId())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(invalidTimeRequest)))
-                        .andExpect(status().isBadRequest())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.INVALID_EXERCISE_TIME.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.INVALID_EXERCISE_TIME.getMessage()));
-            }
-
-            @Test
-            @DisplayName("400 - 과거 날짜로 수정 시 에러를 반환한다")
-            void pastDate() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                ExerciseUpdateDTO.Request pastDateRequest = new ExerciseUpdateDTO.Request(
-                        "2000-01-01", null, null, null, null,
-                        "10:00", "12:00", null, null, null, null
-                );
-
-                mockMvc.perform(patch("/api/exercises/{exerciseId}", exercise.getId())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(pastDateRequest)))
-                        .andExpect(status().isBadRequest())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.PAST_TIME_NOT_ALLOWED.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.PAST_TIME_NOT_ALLOWED.getMessage()));
-            }
-        }
-    }
-
-
-    @Nested
     @DisplayName("POST /api/exercises/{exerciseId}/participants - 운동 신청")
     class JoinExercise {
