+++ src/test/java/umc/cockple/demo/domain/exercise/integration/ExerciseQueryIntegrationTest.java
@@ -93,326 +93,6 @@ class ExerciseQueryIntegrationTest extends IntegrationTestBase {
     }
 
     @Nested
-    @DisplayName("GET /api/exercises/{exerciseId} - 운동 상세 조회")
-    class GetExerciseDetail {
-
-        private Exercise exercise;
-
-        @BeforeEach
-        void setUp() {
-            exercise = exerciseRepository.save(
-                    ExerciseFixture.createExerciseWithAddr(party, LocalDate.now().minusDays(1)));
-        }
-
-        @Nested
-        @DisplayName("성공 케이스")
-        class Success {
-
-            @Test
-            @DisplayName("응답의 모든 주요 필드가 올바르게 반환된다")
-            void 응답의_모든_주요_필드가_올바르게_반환된다() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                Exercise smallExercise = exerciseRepository.save(
-                        ExerciseFixture.createExerciseWithAddr(party, LocalDate.now().minusDays(1), 1));
-
-                memberExerciseRepository.save(MemberFixture.createMemberExercise(normalMember, smallExercise));
-                memberExerciseRepository.save(MemberFixture.createMemberExercise(subManager, smallExercise));
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}", smallExercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.isManager").value(true))
-                        .andExpect(jsonPath("$.data.info.buildingName").value("테스트 체육관"))
-                        .andExpect(jsonPath("$.data.info.location").value("서울특별시 강남구 테헤란로 1"))
-                        .andExpect(jsonPath("$.data.participants.currentParticipantCount").value(1))
-                        .andExpect(jsonPath("$.data.participants.totalCount").value(1))
-                        .andExpect(jsonPath("$.data.participants.manCount").value(1))
-                        .andExpect(jsonPath("$.data.participants.womenCount").value(0))
-                        .andExpect(jsonPath("$.data.participants.list[0].participantNumber").value(1))
-                        .andExpect(jsonPath("$.data.participants.list[0].name").value(normalMember.getMemberName()))
-                        .andExpect(jsonPath("$.data.participants.list[0].gender").value("MALE"))
-                        .andExpect(jsonPath("$.data.participants.list[0].level").isString())
-                        .andExpect(jsonPath("$.data.participants.list[0].participantType").value("PARTY_MEMBER"))
-                        .andExpect(jsonPath("$.data.participants.list[0].partyPosition").value("PARTY_MEMBER"))
-                        .andExpect(jsonPath("$.data.participants.list[0].isWithdrawn").value(false))
-                        .andExpect(jsonPath("$.data.waiting.currentWaitingCount").value(1))
-                        .andExpect(jsonPath("$.data.waiting.manCount").value(0))
-                        .andExpect(jsonPath("$.data.waiting.womenCount").value(1))
-                        .andExpect(jsonPath("$.data.waiting.list[0].name").value(subManager.getMemberName()))
-                        .andExpect(jsonPath("$.data.waiting.list[0].gender").value("FEMALE"))
-                        .andExpect(jsonPath("$.data.waiting.list[0].participantType").value("PARTY_MEMBER"))
-                        .andExpect(jsonPath("$.data.waiting.list[0].partyPosition").value("PARTY_SUBMANAGER"))
-                        .andExpect(jsonPath("$.data.waiting.list[0].isWithdrawn").value(false));
-            }
-
-            @Test
-            @DisplayName("모임장이 조회하면 isManager true로 반환된다")
-            void 모임장이_조회하면_isManager_true로_반환된다() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}", exercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.isManager").value(true));
-            }
-
-            @Test
-            @DisplayName("일반 멤버가 조회하면 isManager false로 반환된다")
-            void 일반_멤버가_조회하면_isManager_false로_반환된다() throws Exception {
-                SecurityContextHelper.setAuthentication(normalMember.getId(), normalMember.getNickname());
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}", exercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.isManager").value(false));
-            }
-
-            @Test
-            @DisplayName("부모임장이 조회해도 isManager false로 반환된다")
-            void 부모임장이_조회해도_isManager_false로_반환된다() throws Exception {
-                SecurityContextHelper.setAuthentication(subManager.getId(), subManager.getNickname());
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}", exercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.isManager").value(false));
-            }
-
-            @Test
-            @DisplayName("모임 외부 회원이 조회해도 isManager false로 반환된다")
-            void 모임_외부_회원이_조회해도_isManager_false로_반환된다() throws Exception {
-                SecurityContextHelper.setAuthentication(outsider.getId(), outsider.getNickname());
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}", exercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.isManager").value(false))
-                        .andExpect(jsonPath("$.data.info.buildingName").value("테스트 체육관"));
-            }
-
-            @Test
-            @DisplayName("정원 초과 참가자는 대기자로 반환된다")
-            void 정원_초과_참가자는_대기자로_반환된다() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                Exercise smallExercise = exerciseRepository.save(
-                        ExerciseFixture.createExerciseWithAddr(party, LocalDate.now().minusDays(1), 1));
-
-                memberExerciseRepository.save(MemberFixture.createMemberExercise(normalMember, smallExercise));
-                memberExerciseRepository.save(MemberFixture.createMemberExercise(subManager, smallExercise));
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}", smallExercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.participants.currentParticipantCount").value(1))
-                        .andExpect(jsonPath("$.data.waiting.currentWaitingCount").value(1));
-            }
-
-            @Test
-            @DisplayName("먼저 가입한 참가자가 더 낮은 participantNumber를 받는다")
-            void 먼저_가입한_참가자가_더_낮은_participantNumber를_받는다() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                memberExerciseRepository.save(MemberFixture.createMemberExercise(normalMember, exercise));
-                memberExerciseRepository.save(MemberFixture.createMemberExercise(subManager, exercise));
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}", exercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.participants.currentParticipantCount").value(2))
-                        .andExpect(jsonPath("$.data.participants.list[0].participantNumber").value(1))
-                        .andExpect(jsonPath("$.data.participants.list[1].participantNumber").value(2))
-                        .andExpect(jsonPath("$.data.participants.list[0].name").value(normalMember.getMemberName()))
-                        .andExpect(jsonPath("$.data.participants.list[1].name").value(subManager.getMemberName()));
-            }
-
-            @Test
-            @DisplayName("참가자의 성별 카운트가 올바르게 반환된다")
-            void 참가자의_성별_카운트가_올바르게_반환된다() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                memberExerciseRepository.save(MemberFixture.createMemberExercise(normalMember, exercise));
-                memberExerciseRepository.save(MemberFixture.createMemberExercise(subManager, exercise));
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}", exercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.participants.currentParticipantCount").value(2))
-                        .andExpect(jsonPath("$.data.participants.manCount").value(1))
-                        .andExpect(jsonPath("$.data.participants.womenCount").value(1));
-            }
-
-            @Test
-            @DisplayName("대기자의 성별 카운트가 올바르게 반환된다")
-            void 대기자의_성별_카운트가_올바르게_반환된다() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                Exercise smallExercise = exerciseRepository.save(
-                        ExerciseFixture.createExerciseWithAddr(party, LocalDate.now().minusDays(1), 1));
-
-                memberExerciseRepository.save(MemberFixture.createMemberExercise(normalMember, smallExercise));
-                memberExerciseRepository.save(MemberFixture.createMemberExercise(subManager, smallExercise));
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}", smallExercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.participants.currentParticipantCount").value(1))
-                        .andExpect(jsonPath("$.data.participants.manCount").value(1))
-                        .andExpect(jsonPath("$.data.participants.womenCount").value(0))
-                        .andExpect(jsonPath("$.data.waiting.currentWaitingCount").value(1))
-                        .andExpect(jsonPath("$.data.waiting.manCount").value(0))
-                        .andExpect(jsonPath("$.data.waiting.womenCount").value(1));
-            }
-
-            @Test
-            @DisplayName("참가자 유형별 partyPosition이 올바르게 반환된다")
-            void 참가자_유형별_partyPosition이_올바르게_반환된다() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                memberExerciseRepository.save(MemberFixture.createMemberExercise(manager, exercise));
-                memberExerciseRepository.save(MemberFixture.createMemberExercise(subManager, exercise));
-                memberExerciseRepository.save(MemberFixture.createMemberExercise(normalMember, exercise));
-                memberExerciseRepository.save(MemberFixture.createExternalMemberExercise(outsider, exercise));
-                guestRepository.save(GuestFixture.createGuest(exercise, manager.getId()));
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}", exercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.participants.list[0].name").value(manager.getMemberName()))
-                        .andExpect(jsonPath("$.data.participants.list[0].participantType").value("PARTY_MEMBER"))
-                        .andExpect(jsonPath("$.data.participants.list[0].partyPosition").value("PARTY_MANAGER"))
-                        .andExpect(jsonPath("$.data.participants.list[1].name").value(subManager.getMemberName()))
-                        .andExpect(jsonPath("$.data.participants.list[1].participantType").value("PARTY_MEMBER"))
-                        .andExpect(jsonPath("$.data.participants.list[1].partyPosition").value("PARTY_SUBMANAGER"))
-                        .andExpect(jsonPath("$.data.participants.list[2].name").value(normalMember.getMemberName()))
-                        .andExpect(jsonPath("$.data.participants.list[2].participantType").value("PARTY_MEMBER"))
-                        .andExpect(jsonPath("$.data.participants.list[2].partyPosition").value("PARTY_MEMBER"))
-                        .andExpect(jsonPath("$.data.participants.list[3].name").value(outsider.getMemberName()))
-                        .andExpect(jsonPath("$.data.participants.list[3].participantType").value("EXTERNAL_PARTICIPANT"))
-                        .andExpect(jsonPath("$.data.participants.list[3].partyPosition").value(nullValue()))
-                        .andExpect(jsonPath("$.data.participants.list[4].name").value("게스트"))
-                        .andExpect(jsonPath("$.data.participants.list[4].participantType").value("GUEST"))
-                        .andExpect(jsonPath("$.data.participants.list[4].partyPosition").value(nullValue()));
-            }
-
-            @Test
-            @DisplayName("게스트 참가자의 inviterName이 반환된다")
-            void 게스트_참가자의_inviterName이_반환된다() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                guestRepository.save(GuestFixture.createGuest(exercise, manager.getId()));
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}", exercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.participants.list[0].participantType").value("GUEST"))
-                        .andExpect(jsonPath("$.data.participants.list[0].inviterName").value(manager.getMemberName()));
-            }
-
-            @Test
-            @DisplayName("활성 회원 참가자는 isWithdrawn false로 반환된다")
-            void 활성_회원_참가자는_isWithdrawn_false로_반환된다() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                memberExerciseRepository.save(MemberFixture.createMemberExercise(normalMember, exercise));
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}", exercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.participants.list[0].isWithdrawn").value(false));
-            }
-
-            @Test
-            @DisplayName("탈퇴 회원 참가자는 isWithdrawn true로 반환된다")
-            void 탈퇴_회원_참가자는_isWithdrawn_true로_반환된다() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                Member withdrawnMember = memberRepository.save(
-                        MemberFixture.createWithdrawnMember("탈퇴회원", "탈퇴닉네임", 8888L));
-
-                memberExerciseRepository.save(MemberFixture.createMemberExercise(withdrawnMember, exercise));
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}", exercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.participants.list[0].isWithdrawn").value(true));
-            }
-        }
-
-        @Nested
-        @DisplayName("실패 케이스")
-        class Failure {
-
-            @Test
-            @DisplayName("존재하지 않는 운동이면 에러를 반환한다")
-            void 존재하지_않는_운동이면_에러를_반환한다() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}", 999L))
-                        .andExpect(status().isNotFound())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.EXERCISE_NOT_FOUND.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.EXERCISE_NOT_FOUND.getMessage()));
-            }
-
-            @Test
-            @DisplayName("존재하지 않는 멤버면 에러를 반환한다")
-            void 존재하지_않는_멤버면_에러를_반환한다() throws Exception {
-                SecurityContextHelper.setAuthentication(999L, "없는멤버");
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}", exercise.getId()))
-                        .andExpect(status().isNotFound())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.MEMBER_NOT_FOUND.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.MEMBER_NOT_FOUND.getMessage()));
-            }
-        }
-    }
-
-    @Nested
-    @DisplayName("GET /api/exercises/{exerciseId}/for-edit - 운동 수정용 상세 조회")
-    class GetExerciseForEdit {
-
-        private Exercise exercise;
-
-        @BeforeEach
-        void setUp() {
-            Exercise exerciseForEdit = ExerciseFixture.createExerciseWithAddr(
-                    party, LocalDate.of(2026, 3, 24), 18);
-            ReflectionTestUtils.setField(exerciseForEdit, "endTime", LocalTime.of(12, 30));
-            ReflectionTestUtils.setField(exerciseForEdit, "notice", "수정 공지사항");
-            exercise = exerciseRepository.save(exerciseForEdit);
-        }
-
-        @Nested
-        @DisplayName("성공 케이스")
-        class Success {
-
-            @Test
-            @DisplayName("응답의 모든 수정용 필드가 올바르게 반환된다")
-            void 응답의_모든_수정용_필드가_올바르게_반환된다() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}/for-edit", exercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.date").value("2026-03-24"))
-                        .andExpect(jsonPath("$.data.buildingName").value("테스트 체육관"))
-                        .andExpect(jsonPath("$.data.roadAddress").value("서울특별시 강남구 테헤란로 1"))
-                        .andExpect(jsonPath("$.data.latitude").value(37.5))
-                        .andExpect(jsonPath("$.data.longitude").value(127.0))
-                        .andExpect(jsonPath("$.data.startTime").value("10:00:00"))
-                        .andExpect(jsonPath("$.data.endTime").value("12:30:00"))
-                        .andExpect(jsonPath("$.data.maxCapacity").value(18))
-                        .andExpect(jsonPath("$.data.allowMemberGuestsInvitation").value(true))
-                        .andExpect(jsonPath("$.data.allowExternalGuests").value(false))
-                        .andExpect(jsonPath("$.data.notice").value("수정 공지사항"));
-            }
-        }
-
-        @Nested
-        @DisplayName("실패 케이스")
-        class Failure {
-
-            @Test
-            @DisplayName("존재하지 않는 운동이면 에러를 반환한다")
-            void 존재하지_않는_운동이면_에러를_반환한다() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}/for-edit", 999L))
-                        .andExpect(status().isNotFound())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.EXERCISE_NOT_FOUND.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.EXERCISE_NOT_FOUND.getMessage()));
-            }
-        }
-    }
-
-    @Nested
     @DisplayName("GET /api/exercises/{exerciseId}/guests - 내가 초대한 운동 게스트 조회")
     class GetMyInvitedGuests {
